Prevent FOUC for the theme button
Rendering based on the value of the dark store, results in a FOUC. Hence, we include both variants of the button and use the `dark:` tailwind variant to add a `display: none` accordingly. Sadly, this means we can no longer use transitions.

Instead, we add a fade animation. Adding the fade animation class directly would cause the buttons to fade during their first render (page load or refresh). So we add the class to the buttons in a click event handler.